Create Confirmation Service

+++ client/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../models/message';
 import { Pagination } from '../models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -16,12 +17,13 @@ export class MessagesComponent implements OnInit {
   pageSize = 5; 
   loading = false; 
 
-  constructor(private messageService: MessageService) { }
+  constructor(private messageService: MessageService, private confirmService: ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
   }
 
+  //Load messages with pagination
   loadMessages() {
     this.loading = true; 
     this.messageService.getMessages(this.pageNumber, this.pageSize, this.container).subscribe(response => {
@@ -31,6 +33,7 @@ export class MessagesComponent implements OnInit {
     })
   }
 
+  //when the pagination page is changed, update the results.
   pageChanged(event: any) {
     if (this.pageNumber !== event.page) {
       this.pageNumber = event.page; 
@@ -38,9 +41,15 @@ export class MessagesComponent implements OnInit {
     }
   }
 
+  //Delete a message.
+  //Confirm service has to be subscribed to anywhere outside of the root guard. 
   deleteMessage(id: number) {
-    this.messageService.deleteMessage(id).subscribe(() => {
-      this.messages.splice(this.messages.findIndex(m => m.id === id), 1); 
+    this.confirmService.confirm('Confirm delete message?', 'This cannot be undone.').subscribe(result => {
+      if (result) {
+        this.messageService.deleteMessage(id).subscribe(() => {
+          this.messages.splice(this.messages.findIndex(m => m.id === id), 1); 
+        })
+      }
     })
   }
 }
